requirements.txt
gspread==3.6.0
python-dotenv==0.15.0
discord.py==1.5.1
python-Levenshtein==0.12.0
google-auth==1.23.0
redis==3.5.3
